Sorry for the noise. I forgot to change the header file.
Subversion: revision:2747

+++ src/measurement/scorep_definitions.h
@@ -187,7 +187,7 @@ SCOREP_DefineSystemTreeNode( SCOREP_SystemTreeNodeHandle parent,
                              const char*                 klass );
 
 SCOREP_LocationGroupHandle
-SCOREP_DefineLocationGroup( uint64_t                    globalLocationGroupId,
+SCOREP_DefineLocationGroup( uint32_t                    globalLocationGroupId,
                             SCOREP_SystemTreeNodeHandle parent,
                             const char*                 name );
